Hoàn thành các chức năng đã được phân công

+++ src/main/java/vn/edu/hcmuaf/fit/sourcedoannoithat/dao/model/Invoice.java
@@ -12,9 +12,9 @@ import java.sql.Timestamp;
 public class Invoice {
     private int id;
     private int orderId;
-    private int total;
+    private double total;
     private int discount;
-    private int finalAmount;
+    private double finalAmount;
     private int paymentMethod;
     private int voucherId;
     private Timestamp issueDate;
